README.md
This was the app I created for my Apple Project Final in my University.

QueueAndPlay is an app for iPhone and Apple tablets that allow one user to host a lobby and allow other users to create a playlist to play on the host device. This primary purpose of this is for party hosts can let other people select what music they wish to hear.

Requirements:

Both Host and guests need to have an internet connection.
The host must designate a name for the lobby and the guests must have the name of that lobby to get in.
The music is provided locally, so the host devices must have music stored in the device.